@@ -1,15 +1,14 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
   <ItemGroup>
+    <FrameworkReference Include="Microsoft.AspNetCore.App" Version="3.1.0" />
     <PackageReference Include="CommandLineParser" Version="2.5.0" />
-    <PackageReference Include="dotenv.net" Version="1.0.6" />
-    <PackageReference Include="Microsoft.Azure.Services.AppAuthentication" Version="1.3.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging" Version="2.2.0" />
-    <PackageReference Include="Microsoft.Identity.Client" Version="4.2.1" />
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.2" />
-    <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="5.5.0" />
+    <!--
+    <PackageReference Include="CasAuth" Version="1.3.0" />
+    -->
+    <ProjectReference Include="../nuget-package/CasAuth.csproj" />
   </ItemGroup>
 </Project>
