@@ -7,6 +7,7 @@ using System.Security.Claims;
 using System.Text.Json;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
+using System.Net.Http;
 
 namespace CasAuth
 {
@@ -53,7 +54,9 @@ namespace CasAuth
                         if (filters.Length < 1) throw new HttpException(404, $"config name of '{name}' is not found.");
 
                         // return the config
-                        var config = await CasConfig.Load(filters);
+                        var httpClientFactory = context.RequestServices.GetService<IHttpClientFactory>();
+                        var httpClient = httpClientFactory.CreateClient("cas");
+                        var config = await CasConfig.Load(httpClient, filters);
                         string json = JsonSerializer.Serialize(config);
                         context.Response.Headers.Add("Content-Type", "application/json; charset=utf-8");
                         await context.Response.WriteAsync(json);
@@ -79,13 +82,12 @@ namespace CasAuth
                     try
                     {
 
-                        // get the identity of the authenticated user
-                        var identity = context.User.Identity as ClaimsIdentity;
-                        if (identity == null || !identity.IsAuthenticated) throw new HttpException(401, "user is not authenticated");
+                        // ensure user is authorized
+                        if (context.User == null) throw new HttpException(401, "user is not authenticated");
 
                         // filter the claims
                         var filter = new string[] { "xsrf", "old", "exp", "iss", "aud" };
-                        var filtered = identity.Claims.ToList();
+                        var filtered = context.User.Claims.ToList();
                         filtered.RemoveAll(c => filter.Contains(c.Type) || c.Type.StartsWith("http://schemas.microsoft.com/"));
 
                         // project to a dictionary
@@ -136,13 +138,9 @@ namespace CasAuth
                     try
                     {
 
-                        // get the identity of the authenticated user
-                        var identity = context.User.Identity as ClaimsIdentity;
-                        if (identity == null || !identity.IsAuthenticated) throw new HttpException(401, "user is not authenticated");
-
-                        // ensure that the user is an admin
-                        if (identity.Claims.Count(c => c.Type == "roles" && c.Value.ToLower() == CasEnv.RoleForAdminFunctions) < 1)
-                            throw new HttpException(403, "user is not authorized");
+                        // ensure user is authorized
+                        if (context.User == null) throw new HttpException(401, "user is not authenticated");
+                        if (!context.User.Claims.HasRole(CasEnv.RoleForAdmin)) throw new HttpException(403, "user is not authorized");
 
                         // find the validator and use it to clear cache
                         var validator = context.RequestServices.GetService<CasTokenValidator>();
@@ -172,13 +170,9 @@ namespace CasAuth
                     try
                     {
 
-                        // get the identity of the authenticated user
-                        var identity = context.User.Identity as ClaimsIdentity;
-                        if (identity == null || !identity.IsAuthenticated) throw new HttpException(401, "user is not authenticated");
-
-                        // ensure that the user is an admin
-                        if (identity.Claims.Count(c => c.Type == "roles" && c.Value.ToLower() == CasEnv.RoleForAdminFunctions) < 1)
-                            throw new HttpException(403, "user is not authorized");
+                        // ensure user is authorized
+                        if (context.User == null) throw new HttpException(401, "user is not authenticated");
+                        if (!context.User.Claims.HasRole(CasEnv.RoleForAdmin)) throw new HttpException(403, "user is not authorized");
 
                         // get the configuration
                         var list = new List<string>();
